phenopipe-web/requirements.txt
alabaster==0.7.10
alembic==0.8.10
APScheduler==3.4.0
Babel==2.5.1
cairocffi==0.7.2
CairoSVG==1.0.22
certifi==2017.11.5
cffi==1.9.1
chardet==3.0.4
click==6.6
cssselect==1.0.0
docutils==0.14
enum==0.4.6
enum34==1.1.6
et-xmlfile==1.0.1
Flask==0.11.1
Flask-Cors==3.0.2
Flask-GraphQL==1.4.0
Flask-JWT-Extended==1.3.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
futures==3.1.1
graphene==1.1.3
graphene-sqlalchemy==1.1.1
graphql-core==1.0.1
graphql-relay==0.4.5
grpcio==1.4.0
grpcio-tools==1.4.0
html5lib==0.999999999
idna==2.6
imagesize==0.7.1
iso8601==0.1.11
itsdangerous==0.24
jdcal==1.3
Jinja2==2.8
lxml==3.6.4
Mako==1.0.6
MarkupSafe==0.23
openpyxl==2.4.8
promise==1.0.1
protobuf==3.4.0
psycopg2==2.6.2
pycparser==2.17
Pygments==2.2.0
PyJWT==1.4.2
Pyphen==0.9.4
python-dateutil==2.6.1
python-editor==1.0.3
python-ldap==2.4.30
pytz==2017.2
redis==2.10.5
requests==2.18.4
rq==0.8.2
simplekv==0.10.0
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.4
tinycss==0.4
typing==3.5.2.2
tzlocal==1.4
urllib3==1.22
WeasyPrint==0.33
webencodings==0.5
Werkzeug==0.11.11
zipstream==1.1.4
